requirements.txt
requests
beautifulsoup4
lxml
pandas
pyarrow
numpy
python-dotenv
